+++ Changelog.md
@@ -6,7 +6,9 @@
 - Fix crash in Preview dialog using 2 instances and shortcut keys (stax76, [#515](/../../issues/515))
 - NVEnc options have a bitrate option (Dendraspis, [#520](/../../issues/520))
 - Blocking assistant warnings are shown in red with Next button disabled (stax76)
-- NVEnc 5.29, QSVEnc 4.13, VCEEnc 6.09 (Win 10 avs Unicode and long path)
+- On systems prior Windows 10 1903 using ANSI characters, avs2pipemod is now used
+  automatically because DJATOM/Asuna/Patman x265 builds have no working ANSI fallback (stax76)
+- NVEnc 5.29, QSVEnc 4.13, VCEEnc 6.09 (Windows 10 avs Unicode and long path)
 - MP4Box 1.1.0 rev506 gcc10.2 Patman
 - Subtitle Edit 3.6.0
 - ffmpeg N-101069 gcc10.2.0 Patman